+++ src/SFManagement.Application/Commands/ChangeTaskStatusCommand.cs
@@ -2,4 +2,4 @@ using SFManagement.Domain.Enums;
 
 namespace SFManagement.Application.Commands;
 
-public record ChangeTaskStatusCommand(int TaskId, ProjectTaskStatus NewStatus);
+public record ChangeTaskStatusCommand(long TaskId, ProjectTaskStatus NewStatus);
