+++ src/SFManagement.Application/DTOs/ProjectDto.cs
@@ -1,6 +1,6 @@
 namespace SFManagement.Application.DTOs;
 
-public record ProjectDto(int Id, string Name, string? DescriptionMd)
+public record ProjectDto(long Id, string Name, string? DescriptionMd)
 {
     public string IdEncrypted { get; init; } = "";
 }
