+++ src/SFManagement.Domain/Entities/Project.cs
@@ -2,11 +2,11 @@ namespace SFManagement.Domain.Entities;
 
 public class Project
 {
-    public int Id { get; init; }
+    public long Id { get; init; }
     public string Name { get; private set; }
     public string? DescriptionMd { get; private set; }
 
-    public Project(int id, string name, string? descriptionMd)
+    public Project(long id, string name, string? descriptionMd)
     {
         Id = id;
         Name = name;
