+++ src/SFManagement.Domain/Entities/SkillCatalogue.cs
@@ -2,12 +2,12 @@ namespace SFManagement.Domain.Entities;
 
 public class SkillCatalogue
 {
-    public int Id { get; init; }
+    public long Id { get; init; }
     public string Name { get; private set; }
     public int VectorPosition { get; init; }
     public bool IsActive { get; private set; }
 
-    public SkillCatalogue(int id, string name, int vectorPosition, bool isActive = true)
+    public SkillCatalogue(long id, string name, int vectorPosition, bool isActive = true)
     {
         Id = id;
         Name = name;
