+++ src/SFManagement.Domain/Entities/TaskAssignment.cs
@@ -2,12 +2,12 @@ namespace SFManagement.Domain.Entities;
 
 public class TaskAssignment
 {
-    public int Id { get; init; }
-    public int TaskId { get; init; }
-    public int WorkerId { get; init; }
+    public long Id { get; init; }
+    public long TaskId { get; init; }
+    public long WorkerId { get; init; }
     public DateTime AssignedAt { get; init; }
 
-    public TaskAssignment(int id, int taskId, int workerId, DateTime assignedAt)
+    public TaskAssignment(long id, long taskId, long workerId, DateTime assignedAt)
     {
         Id = id;
         TaskId = taskId;
